# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aldaro API
  version: "0.1"
servers:
  - url: https://api.aldaro.ai
paths:
...
  /workspaces/launch:
    post:
      summary: Launch workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gpu_type]
              properties:
                gpu_type: { type: string }
                region: { type: string }
      responses:
        "200":
          description: OK
  /workspaces/{id}:
    get:
      summary: Get workspace detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...
  /workspaces/{id}/terminate:
    post:
      summary: Terminate workspace
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...